requirements.txt
pytest
coverage
requests
PyJWT
https://github.com/Ultimaker/CuraPluginOAuth2Module